+++ HISTORY.rst
@@ -3,6 +3,35 @@
 History
 -------
 
+0.8.81 [2023-03-18]
+* fix: fetch_event for multiple events with same name
+
+0.8.80 [2023-03-14]
+* fix web3 for avalanche
+
+0.8.79 [2023-03-14]
+* add arbitrum one token data
+* add optimism token data
+* add avalanche token data
+* add fantom token data
+* fix polygon and bsc native token data
+
+0.8.78 [2023-03-12]
+* support ERC1967Proxy
+
+0.8.77 [2023-03-02]
+* change `force` to `set_loaded` for `set_abi()` and `as_erc20()`
+* support arbitrum one and optimism
+
+0.8.76 [2023-03-01]
+* add network for Optimism, Arbitrum One
+
+0.8.75 [2023-02-27]
+* add staging gateway address
+
+0.8.74 [2023-02-26]
+* make as_erc20() quicker with force=True. Direct injection of ERC20 ABI.
+
 0.8.73 [2023-02-03]
 * fix ledger model setup
 * fix typos in code and doc